新增強化系統介面
- 加入強化系統測試用場景 (未完成)
- 加入UI物件&腳本
- UserDataSample新增示範用的符咒和玉珮文件
- 修復SGJadeEffects變數名稱錯誤的Bug

+++ Assets/EnhanceSys/Scripts/EnhanceIndexer.cs
@@ -21,6 +21,6 @@ public class EnhanceIndexer
 
     static public ShipGear GetShipJadeEff(EnhanceType type, int effectID)
     {
-        return _jadeEffectsSG.Where(effect => effect.EnhType == type).Where(effect => effect.EnhanceID == effectID).ToArray()[0];
+        return _jadeEffectsSG.Where(effect => (int)effect.EnhType == (int)type).Where(effect => effect.EnhanceID == effectID).ToArray()[0];
     }
 }
